+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/session_legacy.php
@@ -1,8 +1,0 @@
-<?php
-
-// To be removed in Symfony 6.0
-$container->loadFromExtension('framework', [
-    'session' => [
-        'handler_id' => null,
-    ],
-]);