+++ src/Symfony/Component/HttpKernel/EventListener/SessionListener.php
@@ -19,10 +19,6 @@ use Symfony\Component\HttpKernel\Event\RequestEvent;
 /**
  * Sets the session in the request.
  *
- * When the passed container contains a "session_storage" entry which
- * holds a NativeSessionStorage instance, the "cookie_secure" option
- * will be set to true whenever the current main request is secure.
- *
  * @author Fabien Potencier <fabien@symfony.com>
  *
  * @final
@@ -34,29 +30,8 @@ class SessionListener extends AbstractSessionListener
         parent::__construct($container, $debug);
     }
 
-    public function onKernelRequest(RequestEvent $event)
-    {
-        parent::onKernelRequest($event);
-
-        if (!$event->isMainRequest() || (!$this->container->has('session') && !$this->container->has('session_factory'))) {
-            return;
-        }
-
-        if ($this->container->has('session_storage')
-            && ($storage = $this->container->get('session_storage')) instanceof NativeSessionStorage
-            && ($mainRequest = $this->container->get('request_stack')->getMainRequest())
-            && $mainRequest->isSecure()
-        ) {
-            $storage->setOptions(['cookie_secure' => true]);
-        }
-    }
-
     protected function getSession(): ?SessionInterface
     {
-        if ($this->container->has('session')) {
-            return $this->container->get('session');
-        }
-
         if ($this->container->has('session_factory')) {
             return $this->container->get('session_factory')->createSession();
         }